backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
beautifulsoup4
openai
python-dotenv